hw5/ArrayList61BTest.java
import org.junit.Test;
import static org.junit.Assert.*;
import java.util.List;

/** ArrayList61BTest. You should write additional tests.
 *  @author Josh Hug
 */

public class ArrayList61BTest {
    @Test
    public void basicTest() {
        List<Integer> L = new ArrayList61B<Integer>();
        L.add(5);
        L.add(10);
        assertTrue(L.contains(5));        
        assertFalse(L.contains(0));
        assertEquals(L.get(0),(Integer) 5);
        assertEquals(L.get(1),(Integer) 10);
        assertEquals(L.size(),2);
        L.add(15);
        assertEquals(L.size(),3);

    }

    /** Runs tests. */
    public static void main(String[] args) {
        jh61b.junit.textui.runClasses(ArrayList61BTest.class);
    }
}   
